requirements.txt
asgiref==3.2.10
Django==3.1.2
django-phone-field==1.8.1
djangorestframework==3.12.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.4.0
